Separate geospatial from non-geospatial tiles
