+++ modules/layers/src/base-tile-layer/utils/tile.js
@@ -1,25 +1,9 @@
-// https://wiki.openstreetmap.org/wiki/Slippy_map_tilenames#Lon..2Flat._to_tile_numbers_2
-export function tile2latLng(x, y, z) {
-  const lng = (x / Math.pow(2, z)) * 360 - 180;
-  const n = Math.PI - (2 * Math.PI * y) / Math.pow(2, z);
-  const lat = (180 / Math.PI) * Math.atan(0.5 * (Math.exp(n) - Math.exp(-n)));
-  return [lng, lat];
-}
-
-export function tile2geoBoundingBox(x, y, z) {
-  const [west, north] = tile2latLng(x, y, z);
-  const [east, south] = tile2latLng(x + 1, y + 1, z);
-  return {west, north, east, south};
-}
-
 export default class Tile {
   constructor({getTileData, x, y, z, onTileLoad, onTileError, tile2boundingBox}) {
     this.x = x;
     this.y = y;
     this.z = z;
-    this.bbox = tile2boundingBox
-      ? tile2boundingBox(this.x, this.y, this.z)
-      : tile2geoBoundingBox(this.x, this.y, this.z);
+    this.bbox = tile2boundingBox(x, y, z);
     this.isVisible = true;
     this.getTileData = getTileData;
     this._data = null;
