+++ modules/layers/src/index.js
@@ -33,6 +33,7 @@ export {default as PolygonLayer} from './polygon-layer/polygon-layer';
 export {default as GeoJsonLayer} from './geojson-layer/geojson-layer';
 export {default as TextLayer} from './text-layer/text-layer';
 export {default as SolidPolygonLayer} from './solid-polygon-layer/solid-polygon-layer';
+export {default as BaseTileLayer} from './base-tile-layer/base-tile-layer';
 
 // Experimental layer exports
 export {default as _MultiIconLayer} from './text-layer/multi-icon-layer/multi-icon-layer';
